fixed __init__
